Чуть упростил код, вынес классы и структуры в отдельные файлы
